Fix looking up line number and colum when there is no exact match.
When a compile error report a colum/error that is not an exact
match in the positionMap dictionary, the last position in the
map with a line number and position before the reported error
should be returned.

The old code failed to update the line number when the line number
matched exactly, sorted the keys by ignored the columns - so it
would try the largest column first (which then obviously failed too),
causing the line before the correct line to be returned.

Much worse, it returned the C# line/col then, and not the corresponding
LSL line/col!